With unittests, integration and functional test, -> next: Documentation

+++ bk_vimist/inventory/models/inventoryItems.py
@@ -7,7 +7,7 @@ class Inventory(TimestampedModel):
     stores inventory items
     '''
     product = models.ForeignKey('inventory.Product', related_name='product_inventory', on_delete=models.PROTECT)
-    quantity = models.SmallIntegerField(validators=[MinValueValidator(0)])
+    quantity = models.DecimalField(max_digits=10, decimal_places=2, validators=[MinValueValidator(0)])
 
     class Meta:
         db_table = 'inventory'
